Remote system db syncronization (#37)
* extract webserver & email logic from Envir

* Fix build after split

* Command for receive systemdb

* syncronization form

* fix load partial data
